Compiler warnings

+++ BugsFramework/Source/Extensions/NSArray+Shuffling.m
@@ -14,7 +14,7 @@
 - (NSArray *)shuffledArray;
 {
 	NSMutableArray *shuffled = [NSMutableArray arrayWithArray:self];
-	int count = [shuffled count];
+	NSUInteger count = [shuffled count];
 	for (int i = 0; i < count; i++) {
 		[shuffled exchangeObjectAtIndex:i withObjectAtIndex:(random() % count)];
 	}
